refactor(mm-next): move gql operations to /membership
