--- packages/mirror-media-next/apollo/profile/query/fetch-member-profile.js
@@ -1,19 +1,0 @@
-import { gql } from '@apollo/client'
-
-export const fetchMemberProfileByFirebaseId = gql`
-  query fetchMemberProfile($firebaseId: String!) {
-    member(where: { firebaseId: $firebaseId }) {
-      id
-      firebaseId
-      email
-      name
-      gender
-      birthday
-      phone
-      country
-      city
-      district
-      address
-    }
-  }
-`
