+++ packages/mirror-media-next/components/profile/user-profile-form.js
@@ -7,7 +7,7 @@ import PrimaryButton from '../shared/buttons/primary-button'
 import DefaultButton from '../shared/buttons/default-button'
 import { useRouter } from 'next/router'
 import client from '../../apollo/apollo-client'
-import { updateMember } from '../../apollo/profile/mutation/update-member-profile'
+import { updateMember } from '../../apollo/membership/mutation/member'
 import { useMembership } from '../../context/membership'
 import dayjs from 'dayjs'
 import utc from 'dayjs/plugin/utc'
